Allow filenames including concept dir names
The way it works now is that if you have a directory in `app/concepts`
that contains a the name of a concept directory from the excludes list

```
%w{ callback cell contract operation policy representer view }

```

then it is excluded from the loader

I had for example a directory `app/concepts/reviews` which I imagine is
pretty common and it was never getting included and it was driving me
crazy >:(

This is in inelegant solution. Test is passing but needs to be prettier.